ml_api/requirements.txt
annotated-types==0.6.0
appscript==1.2.5
asgiref==3.7.2
attrs==23.2.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.5
cymem==2.0.8
distlib==0.3.8
Django==5.0.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
docx==0.2.4
docx2pdf==0.1.8
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
filelock==3.13.1
Flask==3.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langcodes==3.3.0
lxml==5.3.1
MarkupSafe==2.1.5
murmurhash==1.0.10
mysql-connector==2.2.9
mysql-connector-python==8.3.0
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
platformdirs==4.2.0
preshed==3.0.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
PyMySQL==1.1.0
pypdfium2==4.29.0
pyresparser==1.0.6
pyrsistent==0.20.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
referencing==0.34.0
regex==2023.12.25
reportlab==4.2.2
requests==2.31.0
rpds-py==0.18.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sortedcontainers==2.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.1.3
xlwt==1.3.0
